Create the live-search-maxlength parameter
Closes mauriciogentile/angular-livesearch#7.

Closes 18F/hub#119.

As mentioned in the above issues, the nine-character query string limit was
problematic for our use case. This effectively removes the limit, but users
can elect to limit the maximum query string limit by setting the
`live-search-maxlength` attribute.

Also fixes the `should not invoke search callback if input length is less than
3` test spec, as I could not make the test fail until applying the updates
from this commit.